Make Paperclip and Webmock spec support config files reusable
Move to `pageflow-support` gem. The Webmock config is coupled to the
Capybara webdriver setup. Paperclip config is required to clear
attachments created during tests.

REDMINE-17301
